+++ examples/Download/src/downloadVideoAndAudio.js
@@ -1,7 +1,9 @@
 const fs = require('fs'),
     { YtdlCore } = require('@ybd-project/ytdl-core'),
     ytdl = new YtdlCore({
-        lang: 'en',
+        hl: 'en',
+        gl: 'US',
+        streamType: 'nodejs',
     });
 
 // Video: Never Gonna give you up
@@ -10,4 +12,8 @@ const VIDEO_ID = 'dQw4w9WgXcQ';
 /* Normal usage (Full Info) */
 ytdl.download(`https://www.youtube.com/watch?v=${VIDEO_ID}`, {
     filter: 'videoandaudio',
-}).pipe(fs.createWriteStream(`./${VIDEO_ID}.mp3`));
+}).then((stream) => {
+    stream.pipe(fs.createWriteStream(`./${VIDEO_ID}.mp4`));
+}).catch((err) => {
+    console.error(err);
+});