+++ examples/Download/src/downloadVideoAndAudio.ts
@@ -1,14 +1,20 @@
-import { YtdlCore } from '@ybd-project/ytdl-core';
+import { YtdlCore, YTDL_NodejsStreamType } from '@ybd-project/ytdl-core';
 import fs from 'fs';
 
 const ytdl = new YtdlCore({
-    lang: 'en',
+    hl: 'en',
+    gl: 'US',
+    streamType: 'nodejs',
 });
 
 // Video: Never Gonna give you up
 const VIDEO_ID = 'dQw4w9WgXcQ';
 
 /* Normal usage (Full Info) */
-ytdl.download(`https://www.youtube.com/watch?v=${VIDEO_ID}`, {
+ytdl.download<YTDL_NodejsStreamType>(`https://www.youtube.com/watch?v=${VIDEO_ID}`, {
     filter: 'videoandaudio',
-}).pipe(fs.createWriteStream(`./${VIDEO_ID}.mp3`));
+}).then((stream) => {
+    stream.pipe(fs.createWriteStream(`./${VIDEO_ID}.mp4`));
+}).catch((err) => {
+    console.error(err);
+});