+++ src/platforms/default/Default.ts
@@ -1,6 +1,7 @@
 import fs from 'fs';
 import path from 'path';
 import os from 'os';
+import { Readable } from 'stream';
 
 import { Platform } from '@/platforms/Platform';
 import { CacheWithMap, YtdlCore_Cache } from '@/platforms/utils/Classes';
@@ -13,7 +14,7 @@ import('./PoToken.mjs').then((m) => {
     const SHIM = Platform.getShim();
     SHIM.poToken = m.generatePoToken;
     Platform.load(SHIM);
-})
+});
 
 class FileCache implements YtdlCore_Cache {
     private timeouts: Map<string, NodeJS.Timeout> = new Map();
@@ -156,8 +157,8 @@ Platform.load({
     fileCache: new FileCache(),
     fetcher: (url, options) => fetch(url, options),
     poToken: () => Promise.resolve({ poToken: '', visitorData: '' }),
-    default: {
-        options: {
+    options: {
+        download: {
             hl: 'en',
             gl: 'US',
             includesPlayerAPIResponse: false,
@@ -168,6 +169,10 @@ Platform.load({
             disableDefaultClients: false,
             disableFileCache: false,
             parsesHLSFormat: true,
+        },
+        other: {
+            logDisplay: ['info', 'success', 'warning', 'error'],
+            noUpdate: false,
         },
     },
     requestRelated: {
@@ -184,40 +189,6 @@ Platform.load({
 });
 
 import { YtdlCore } from '@/YtdlCore';
-
-YtdlCore.writeStreamToFile = async function (readableStream: ReadableStream, filePath: string) {
-    return new Promise((resolve, reject) => {
-        const WRITE_STREAM = fs.createWriteStream(filePath);
-
-        async function pump() {
-            const READER = readableStream.getReader();
-            try {
-                while (true) {
-                    const { done, value } = await READER.read();
-
-                    if (done) {
-                        WRITE_STREAM.end();
-                        break;
-                    }
-
-                    if (!WRITE_STREAM.write(Buffer.from(value))) {
-                        await new Promise((resolve) => WRITE_STREAM.once('drain', resolve));
-                    }
-                }
-            } catch (err: any) {
-                WRITE_STREAM.destroy(err);
-                reject(err);
-            } finally {
-                READER.releaseLock();
-            }
-        }
-
-        pump();
-
-        WRITE_STREAM.on('finish', resolve);
-        WRITE_STREAM.on('error', reject);
-    });
-};
 
 export * from '@/types/index';
 export { YtdlCore };